Refactor to externalise common launch code
